+++ app/src/main/java/com/github/gzuliyujiang/fallback/activity/OptionPickerFragment.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2016-present 贵州纳雍穿青人李裕江<1032694760@qq.com>
+ *
+ * The software is licensed under the Mulan PSL v2.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *     http://license.coscl.org.cn/MulanPSL2
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR FIT FOR A PARTICULAR
+ * PURPOSE.
+ * See the Mulan PSL v2 for more details.
+ */
+
+package com.github.gzuliyujiang.fallback.activity;
+
+import android.app.Dialog;
+import android.os.Bundle;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+
+import com.github.gzuliyujiang.fallback.R;
+import com.github.gzuliyujiang.wheelpicker.OptionPicker;
+import com.github.gzuliyujiang.wheelpicker.contract.OnOptionPickedListener;
+
+/**
+ * {@link androidx.fragment.app.Fragment}形式的弹窗
+ *
+ * @author 贵州山野羡民（1032694760@qq.com）
+ * @since 2022/11/04
+ */
+public class OptionPickerFragment extends DialogFragment {
+
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+        OptionPicker picker = new OptionPicker(requireActivity());
+        picker.setData("土人", "里民子", "羡民", "穿青人", "不在56个民族之内", "未定民族");
+        picker.setOnOptionPickedListener(new OnOptionPickedListener() {
+            @Override
+            public void onOptionPicked(int position, Object item) {
+                Toast.makeText(requireContext(), item.toString(), Toast.LENGTH_SHORT).show();
+            }
+        });
+        picker.getWheelView().setStyle(R.style.WheelStyleDemo);
+        return picker;
+    }
+
+}